Como sugerido o controller não conheci agora a entity, foi criado um service que intermedia os 2
