+++ src/main/kotlin/br/com/estudante/customerapi/controllers/CustomerController.kt
@@ -1,10 +1,9 @@
 package br.com.estudante.customerapi.controllers
 
-import br.com.estudante.customerapi.entity.CustomerEntity
 import br.com.estudante.customerapi.exceptions.CustomerException
-import br.com.estudante.customerapi.repository.CustomerRepository
 import br.com.estudante.customerapi.rest.CustomerRequest
 import br.com.estudante.customerapi.rest.CustomerResponse
+import br.com.estudante.customerapi.service.CustomerService
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.http.HttpStatus
 import org.springframework.http.ResponseEntity
@@ -16,7 +15,6 @@ import org.springframework.web.bind.annotation.RequestBody
 import org.springframework.web.bind.annotation.RequestMapping
 import org.springframework.web.bind.annotation.RestController
 import org.springframework.web.context.request.WebRequest
-import java.util.*
 import javax.validation.Valid
 
 @RestController
@@ -24,19 +22,11 @@ import javax.validation.Valid
 class CustomerController {
 
   @Autowired
-  lateinit var repository: CustomerRepository
+  lateinit var service : CustomerService
 
  @PostMapping
- fun create(@RequestBody @Valid customerRequest: CustomerRequest) : ResponseEntity<CustomerResponse> {
-   val newCustomer = CustomerEntity(
-     UUID.randomUUID().toString(),
-     customerRequest.name,
-     customerRequest.personCode,
-     customerRequest.postalCode,
-     customerRequest.email
-   )
-
-   return ResponseEntity.status(HttpStatus.CREATED).body(CustomerResponse(repository.save(newCustomer).id))
+ fun create(@RequestBody @Valid customerRequestData: CustomerRequest) : ResponseEntity<CustomerResponse> {
+   return service.toCustomerEntity(customerRequestData)
  }
 
  @ExceptionHandler(MethodArgumentNotValidException::class)
